feat: move 'match preparation' status to a dialog
