+++ src/pages/GamePage/AwaitingView/PrepareMatchDialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PrepareMatchDialog'
