+++ src/pages/GamePage/GamePage.tsx
@@ -16,7 +16,6 @@ import PlayersBettingView from './PlayersBettingView'
 import PlayersJoiningView from './PlayersJoiningView'
 import PlayingView from './PlayingView'
 import ReportingHitsView from './ReportingHitsView'
-import SettingUpMatchView from './SettingUpMatchView'
 
 const viewByStatusMap: Record<GameStatus, FC<{ game: Game }>> = {
   [GameStatus.AWAITING]: AwaitingView,
@@ -25,7 +24,6 @@ const viewByStatusMap: Record<GameStatus, FC<{ game: Game }>> = {
   [GameStatus.PLAYERS_JOINING]: PlayersJoiningView,
   [GameStatus.PLAYING]: PlayingView,
   [GameStatus.REPORTING_HITS]: ReportingHitsView,
-  [GameStatus.SETTING_UP_MATCH]: SettingUpMatchView,
 }
 
 function GamePage() {
@@ -35,7 +33,7 @@ function GamePage() {
   const ActiveView = useMemo(() => {
     return activeGame ? viewByStatusMap[activeGame.status] : () => null
   }, [activeGame?.status]) // eslint-disable-line react-hooks/exhaustive-deps
-  const smallViews = [GameStatus.AWAITING, GameStatus.CLOSED, GameStatus.SETTING_UP_MATCH]
+  const smallViews = [GameStatus.AWAITING, GameStatus.CLOSED]
 
   function handleChangeStatus(event: ChangeEvent<HTMLInputElement>) {
     updateGame({ status: event.target.value as GameStatus })
@@ -58,9 +56,8 @@ function GamePage() {
             value={activeGame?.status}
             onChange={handleChangeStatus}
           >
-            <MenuItem value={GameStatus.AWAITING}>Aguardando</MenuItem>
             <MenuItem value={GameStatus.PLAYERS_JOINING}>Adicionar jogadores</MenuItem>
-            <MenuItem value={GameStatus.SETTING_UP_MATCH}>Preparar partida</MenuItem>
+            <MenuItem value={GameStatus.AWAITING}>Aguardando</MenuItem>
             <MenuItem value={GameStatus.PLAYERS_BETTING}>Registrar apostas</MenuItem>
             <MenuItem value={GameStatus.PLAYING}>Em jogo</MenuItem>
             <MenuItem value={GameStatus.REPORTING_HITS}>Registrar resultados</MenuItem>
