--- src/pages/GamePage/SettingUpMatchView/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SettingUpMatchView'
